+++ src/Connection.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Phlib\Beanstalk;
 
 use Phlib\Beanstalk\Connection\Socket;
-use Phlib\Beanstalk\Connection\SocketInterface;
 use Phlib\Beanstalk\Exception\InvalidArgumentException;
 use Phlib\Beanstalk\Exception\NotFoundException;
 
@@ -18,7 +17,7 @@ class Connection implements ConnectionInterface
 
     private string $name;
 
-    private SocketInterface $socket;
+    private Socket $socket;
 
     private string $using = self::DEFAULT_TUBE;